move llm_chat js mode changesl to llm_thread module
